Photocast integration
The plugin is now able to add user's albums in photocast list.
Private albums will be available for 2 weeks, a refresh to
update the token will be needed after this period.
For now, the protected albums can't be accessed with this token,
so the plugin won't add them in the list.